mysuretyplus/backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.14.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
tqdm==4.66.1
python-decouple==3.8
redis==5.0.1
celery==5.3.4
daphne==4.0.0
